test(gc): split box release tests out of box.rs; allowlist their header cast
box.rs reached 2228 lines, over the 2000-line cap. Moved mod release_tests
into box/release_tests.rs and allowlisted its GcHeader cast: the closure is
allocated by the test and it needs a MUTABLE header to toggle GC_FLAG_MARKED,
which try_read_gc_header's shared ref cannot express.
